requirements.txt
hydra-core
hydra-colorlog
torch
torchvision
pytorch-lightning
gmr
scikit-learn
numpy
gym
tqdm
scipy
imageio[ffmpeg]
termcolor
pymanopt
mayavi
wandb
lightning-bolts